Move editor styles and theme to its own file and folder
