+++ src/stories/index.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { dark, light } from "../theme";
+import { dark, light } from "../styles/theme";
 import Editor from "..";
 
 const docSearchResults = [
